refactor(commits): move commit_attribution from Audit to Users
`commit_attribution` is fundamentally a user-resolution operation that
happens to read audit context — it belongs as a method on `Users` rather
than as an associated fn of `Audit` taking a `&Users` parameter.

Call sites become `users.commit_attribution()` instead of the awkward
`Audit::commit_attribution(&users)` form.
